EmberScopeProvider: Rename to EmberModuleReferenceContributor
