+++ src/main/resources/META-INF/plugin.xml
@@ -66,7 +66,7 @@
     </extensions>
 
     <extensions defaultExtensionNs="JavaScript">
-        <moduleReferenceContributor implementation="com.emberjs.resolver.EmberAppReferenceContributor"/>
+        <moduleReferenceContributor implementation="com.emberjs.resolver.EmberModuleReferenceContributor"/>
     </extensions>
 
     <actions>
